requirements.txt
MouseInfo==0.0.4
Pillow==6.1.0
PyAutoGUI==0.9.47
PyGetWindow==0.0.7
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
python-dotenv==0.10.3
PyTweening==1.0.3
